modeltraining/dataset/requirements.txt
numpy
pandas
matplotlib
opencv-python
pillow
lxml
xmltodict
jupyter
torch
torchvision
fastapi
uvicorn